+++ calendarview/src/main/java/com/haibin/calendarview/BaseView.java
@@ -112,12 +112,12 @@ public abstract class BaseView extends View implements View.OnClickListener, Vie
     /**
      * 每一项的高度
      */
-    public int mItemHeight;
+    public float mItemHeight;
 
     /**
      * 每一项的宽度
      */
-    public int mItemWidth;
+    public float mItemWidth;
 
     /**
      * Text的基线
@@ -459,12 +459,12 @@ public abstract class BaseView extends View implements View.OnClickListener, Vie
     }
 
 
-    protected int getCalendarPaddingLeft() {
+    protected float getCalendarPaddingLeft() {
         return mDelegate != null ? mDelegate.getCalendarPaddingLeft() : 0;
     }
 
 
-    protected int getCalendarPaddingRight() {
+    protected float getCalendarPaddingRight() {
         return mDelegate != null ? mDelegate.getCalendarPaddingRight() : 0;
     }
