+++ calendarview/src/main/java/com/haibin/calendarview/RangeMonthView.java
@@ -71,8 +71,8 @@ public abstract class RangeMonthView extends BaseMonthView {
      * @param j        j
      */
     protected void draw(Canvas canvas, Calendar calendar, int calendarIndex, int i, int j) {
-        int x = j * mItemWidth + mDelegate.getCalendarPaddingLeft();
-        int y = i * mItemHeight;
+        float x = j * mItemWidth + mDelegate.getCalendarPaddingLeft();
+        float y = i * mItemHeight;
         onLoopStart(x, y);
         boolean isSelected = isCalendarSelected(calendar);
         boolean hasScheme = calendar.hasScheme();
@@ -267,7 +267,7 @@ public abstract class RangeMonthView extends BaseMonthView {
      * @param isSelectedNext 下一个日期是否选中
      * @return 是否继续绘制onDrawScheme，true or false
      */
-    protected abstract boolean onDrawSelected(Canvas canvas, Calendar calendar, int x, int y, boolean hasScheme,
+    protected abstract boolean onDrawSelected(Canvas canvas, Calendar calendar, float x, float y, boolean hasScheme,
                                               boolean isSelectedPre, boolean isSelectedNext);
 
     /**
@@ -279,7 +279,7 @@ public abstract class RangeMonthView extends BaseMonthView {
      * @param y          日历Card y起点坐标
      * @param isSelected 是否选中
      */
-    protected abstract void onDrawScheme(Canvas canvas, Calendar calendar, int x, int y, boolean isSelected);
+    protected abstract void onDrawScheme(Canvas canvas, Calendar calendar, float x, float y, boolean isSelected);
 
 
     /**
@@ -292,5 +292,5 @@ public abstract class RangeMonthView extends BaseMonthView {
      * @param hasScheme  是否是标记的日期
      * @param isSelected 是否选中
      */
-    protected abstract void onDrawText(Canvas canvas, Calendar calendar, int x, int y, boolean hasScheme, boolean isSelected);
+    protected abstract void onDrawText(Canvas canvas, Calendar calendar, float x, float y, boolean hasScheme, boolean isSelected);
 }
